requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.9.0
astor==0.8.1
beautifulsoup4==4.13.4
brotlicffi @ file:///C:/b/abs_f4lixvzmxt/croot/brotlicffi_1736183286104/work
certifi @ file:///C:/b/abs_43o9lm92p9/croot/certifi_1750863864870/work/certifi
cffi @ file:///C:/b/abs_29_b57if3f/croot/cffi_1736184144340/work
chardet==5.2.0
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
decorator==5.2.1
defusedxml==0.7.1
easyocr==1.7.2
exceptiongroup==1.3.0
fastapi==0.115.14
filelock==3.13.1
fire==0.7.0
flatbuffers==25.2.10
fonttools==4.58.4
fsspec==2024.6.1
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.4
kiwisolver==1.4.8
langdetect==1.0.9
lazy_loader==0.4
lmdb==1.6.2
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
networkx==3.3
ninja==1.11.1.4
numpy==1.26.4
onnxruntime-gpu==1.22.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
packaging==25.0
paddleocr==2.10.0
paddlepaddle-gpu==2.6.2
pandas==2.3.0
pillow==10.3.0
platformdirs==4.2.2
protobuf==3.20.0
psutil==6.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.9.2
pydantic-settings==2.3.4
pydantic_core==2.23.4
Pygments==2.19.2
pyparsing==3.2.3
pypdfium2==4.30.1
pyreadline3==3.5.4
PySocks @ file:///C:/ci_311/pysocks_1676425991111/work
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-doctr==0.12.0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.1
RapidFuzz==3.13.0
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
safetensors==0.4.3
scikit-image==0.25.2
scipy==1.14.0
seaborn==0.13.2
shapely==2.1.1
shellingham==1.5.4
simsimd==6.4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.40.0
stringzilla==3.12.5
sympy==1.12.1
tabulate==0.9.0
termcolor==3.1.0
tifffile==2025.6.11
timm==0.9.16
tokenizers==0.19.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.66.4
typer==0.12.3
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
tzdata==2025.2
ultralytics==8.2.53
ultralytics-thop==2.0.14
urllib3==2.2.1
uvicorn==0.30.1
validators==0.35.0
wcwidth==0.2.13
websockets==12.0
win-inet-pton @ file:///C:/ci_311/win_inet_pton_1676425458225/work
xxhash==3.4.1
